HBSD: change vfs_setopt{,_part}(...)'s len type, to check boundaries, and follow bcopy(...)'s expectation
FreeBSD-PR: 196474
(cherry picked from commit 7cc6ffbda459280327491d43cf4514f7296728d4)

+++ sys/kern/vfs_mount.c
@@ -1617,7 +1617,7 @@ vfs_scanopt(struct vfsoptlist *opts, const char *name, const char *fmt, ...)
 }
 
 int
-vfs_setopt(struct vfsoptlist *opts, const char *name, void *value, int len)
+vfs_setopt(struct vfsoptlist *opts, const char *name, void *value, size_t len)
 {
 	struct vfsopt *opt;
 
@@ -1638,7 +1638,7 @@ vfs_setopt(struct vfsoptlist *opts, const char *name, void *value, int len)
 }
 
 int
-vfs_setopt_part(struct vfsoptlist *opts, const char *name, void *value, int len)
+vfs_setopt_part(struct vfsoptlist *opts, const char *name, void *value, size_t len)
 {
 	struct vfsopt *opt;
